requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
tldextract>=5.1.2
python-dotenv>=1.0.1
charset-normalizer>=3.3.2
dnspython>=2.6.1
email-validator>=2.2.0
google-generativeai>=0.8.3
